Need ffi_set for mutable structs, even if fields isn't publicly writable  a field of a struct can be mutated by a mutating method
Given a swift struct like:

/// <!-- FishyJoes.export(Counter) -->
struct Counter {
    public private(set) var count: Int = 0
    /// <!-- FishyJoes.export(tick) -->
    mutating func tick() { count += 1 }
}
The real "count" variable lives on the dart side, but the "tick" function lives on the swift side. To call "tick" from dart

count is read from the dart "Counter" object
a new swift "Counter" is created with with same count
The mutating method "tick" is then run on the swift counter
The changes are copied back to the original dart "Counter"
In order to do step 4, an ffi setter is needed for count. Even though it's not publicly writable.