feat: get extraDynamicLibrariesToInstall and use for installLibrary
extraDynamicLibsToInstall is different from config.extraDynamicLibraries because it includes the extraDynamicLibraries of all the dependencies as well, instead of the ones just for this module. This is needed because say if CriSVG depends on the dynamic library CanvasModel and CriCanvas also depends on CanvasModel and on CriSVG as well, then we must install the CanvasModel library i.e. copy over the dynamic library to where the language environment needs it to run and test with, but we don't want to pass it to the generate/sourcery step where it's used to generate the kotlin LoaderRepresentative.kt file, because CriSVG will already load it there, so if we try to load the same dynamic CanvasModel library in CriCanvas's LoaderRepresentative we will get an error on Windows and Linux.